Fixed the problem with the vast_pdf library not being found without breaking other things.
